Server/Containers/api/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib[bcrypt]
pyjwt
python-multipart
